src/server/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
pinecone
langchain-pinecone
tiktoken
pydantic
python-multipart
pypdf
protobuf
langchain-community
duckduckgo-search
ddgs
python-docx
docx2txt
langchainhub
